% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggingbwsel-package.R
\docType{package}
\name{baggingbwsel-package}
\alias{baggingbwsel}
\alias{baggingbwsel-package}
\title{baggingbwsel: Bagging bandwidth selection in kernel density and 
regression estimation}
\description{
This package implements bagging bandwidth selection methods for the 
Parzen-Rosenblatt kernel density estimator, and for the Nadaraya-Watson and 
local polynomial kernel regression estimators. 
These bandwidth selectors can achieve greater statistical precision than their 
non-bagged counterparts while being computationally fast. 
See Barreiro-Ures et al. (2021a) and Barreiro-Ures et al. (2021b).
}
\references{
Barreiro-Ures, D., Cao, R., Francisco-Fernández, M., & Hart, J. D. (2021a). 
Bagging cross-validated bandwidths with application to big data. 
\emph{Biometrika}, \bold{108}(4), 981-988, \doi{10.1093/biomet/asaa092}.

Barreiro-Ures, D., Cao, R., & Francisco-Fernández, M. (2021b).  
Bagging cross-validated bandwidth selection in nonparametric regression 
estimation with applications to large-sized samples.
\emph{arXiv preprint}, \url{https://arxiv.org/abs/2105.04134}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rubenfcasal.github.io/baggingbwsel/}
  \item \url{https://github.com/rubenfcasal/baggingbwsel/}
  \item Report bugs at \url{https://github.com/rubenfcasal/baggingbwsel/issues/}
}

}
\author{
\strong{Maintainer}: Ruben Fernandez-Casal \email{rubenfcasal@gmail.com}

Authors:
\itemize{
  \item Daniel Barreiro-Ures \email{daniel.barreiro.ures@udc.es}
  \item Jeffrey Hart
  \item Ricardo Cao
  \item Mario Francisco-Fernandez
}

}
